+++ app/src/main/java/com/haretskiy/pavel/calculatofortests/calculator/Calculator.kt
@@ -2,5 +2,5 @@ package com.haretskiy.pavel.calculatofortests.calculator
 
 interface Calculator {
 
-    fun calculate(a: Int?, b: Int?, operation: (a: Int?, b: Int?) -> Int? ): Int?
+    fun calculate(a: Float?, b: Float?, operation: (a: Float?, b: Float?) -> Float? ): Float?
 }
