+++ app/src/main/java/com/haretskiy/pavel/calculatofortests/calculator/CalculatorImpl.kt
@@ -2,6 +2,6 @@ package com.haretskiy.pavel.calculatofortests.calculator
 
 class CalculatorImpl : Calculator {
 
-    override fun calculate(a: Int?, b: Int?, operation: (a: Int?, b: Int?) -> Int?) = operation(a, b)
+    override fun calculate(a: Float?, b: Float?, operation: (a: Float?, b: Float?) -> Float?) = operation(a, b)
 
 }
